requirements.txt
numpy==1.25.0
streamlit==1.24.0
torch==2.0.1
torchvision==0.15.2
openai==0.27.8
Pillow==9.0.1